requirements.txt
# Base requirements
loguru

# Image requirements
numpy
Pillow
opencv-python

# ML requirements
transform
tensorflow
